chore(bazel): Bazelify chart component.
