+++ libs/barista-components/chart/BUILD.bazel
@@ -1,4 +1,6 @@
-load("//tools/bazel_rules:index.bzl", "stylelint")
+load("@io_bazel_rules_sass//:defs.bzl", "multi_sass_binary")
+load("@npm//@bazel/typescript:index.bzl", "ts_config")
+load("//tools/bazel_rules:index.bzl", "jest", "ng_module_view_engine", "stylelint")
 
 package(default_visibility = ["//visibility:public"])
 
@@ -16,7 +18,97 @@ filegroup(
     ]),
 )
 
+ng_module_view_engine(
+    name = "compile",
+    srcs = glob(
+        include = ["**/*.ts"],
+        exclude = [
+            "**/*.spec.ts",
+            "src/test-setup.ts",
+        ],
+    ),
+    angular_assets = [
+        ":styles",
+        ":src/chart.html",
+        ":src/heatfield/chart-heatfield.html",
+        ":src/range/range.html",
+        ":src/selection-area/selection-area.html",
+        ":src/timestamp/timestamp.html",
+    ],
+    module_name = "@dynatrace/barista-components/chart",
+    tsconfig = "tsconfig_lib",
+    deps = [
+        "//libs/barista-components/button:compile",
+        "//libs/barista-components/core:compile",
+        "//libs/barista-components/formatters:compile",
+        "//libs/barista-components/icon:compile",
+        "//libs/barista-components/loading-distractor:compile",
+        "//libs/barista-components/theming:compile",
+        "@npm//@angular/cdk",
+        "@npm//@angular/common",
+        "@npm//@angular/core",
+        "@npm//highcharts",
+        "@npm//lodash-es",
+        "@npm//rxjs",
+    ],
+)
+
+multi_sass_binary(
+    name = "styles",
+    srcs = [
+        "src/chart.scss",
+        "src/heatfield/_chart-heatfield-theme.scss",
+        "src/heatfield/chart-heatfield.scss",
+        "src/range/range.scss",
+        "src/selection-area/selection-area.scss",
+        "src/timestamp/timestamp.scss",
+        "src/tooltip/chart-tooltip.scss",
+    ],
+)
+
 stylelint(
     name = "stylelint",
     srcs = glob(["**/*.scss"]),
+)
+
+jest(
+    name = "test",
+    srcs = glob(include = ["**/*.spec.ts"]),
+    jest_config = ":jest.config.json",
+    setup_file = ":src/test-setup.ts",
+    ts_config = ":tsconfig_test",
+    deps = [
+        ":compile",
+        "//libs/barista-components/core:compile",
+        "//libs/barista-components/formatters:compile",
+        "//libs/barista-components/icon:compile",
+        "//libs/barista-components/key-value-list:compile",
+        "//libs/barista-components/overlay:compile",
+        "//libs/barista-components/theming:compile",
+        "//libs/testing/browser",
+        "@npm//@angular/cdk",
+        "@npm//@angular/common",
+        "@npm//@angular/core",
+        "@npm//@angular/platform-browser",
+        "@npm//highcharts",
+        "@npm//rxjs",
+    ],
+)
+
+ts_config(
+    name = "tsconfig_lib",
+    src = "tsconfig.lib.json",
+    deps = [
+        "tsconfig.json",
+        "//libs/barista-components:tsconfig",
+    ],
+)
+
+ts_config(
+    name = "tsconfig_test",
+    src = "tsconfig.spec.json",
+    deps = [
+        "tsconfig.json",
+        "//libs/barista-components:tsconfig",
+    ],
 )
