+++ libs/barista-components/chart/jest.config.json
@@ -1,0 +1,8 @@
+{
+  "name": "chart",
+  "snapshotSerializers": [
+    "jest-preset-angular/build/AngularNoNgAttributesSnapshotSerializer.js",
+    "jest-preset-angular/build/AngularSnapshotSerializer.js",
+    "jest-preset-angular/build/HTMLCommentSerializer.js"
+  ]
+}
