+++ libs/barista-components/chart/tsconfig.lib.json
@@ -8,13 +8,5 @@
     "types": [],
     "lib": ["dom", "es2018"]
   },
-  "angularCompilerOptions": {
-    "annotateForClosureCompiler": true,
-    "skipTemplateCodegen": true,
-    "strictMetadataEmit": true,
-    "fullTemplateTypeCheck": true,
-    "strictInjectionParameters": true,
-    "enableResourceInlining": true
-  },
   "exclude": ["src/test-setup.ts", "**/*.spec.ts"]
 }
