+++ package.json
@@ -124,7 +124,7 @@
     "@types/express": "^4.17.2",
     "@types/glob": "^7.1.1",
     "@types/inquirer": "^6.5.0",
-    "@types/jest": "^26.0.0",
+    "@types/jest": "^26.0.5",
     "@types/lodash": "^4.14.155",
     "@types/lodash-es": "^4.17.3",
     "@types/markdown-it": "^10.0.1",
@@ -153,7 +153,7 @@
     "highcharts": "^7.2.1",
     "husky": "^4.2.5",
     "inquirer": "^7.2.0",
-    "jest": "^26.0.1",
+    "jest": "^26.0.5",
     "jest-junit": "^10.0.0",
     "jest-preset-angular": "^8.2.1",
     "jest-serializer-path": "^0.1.15",
